+++ Services/CMdm.Services/CustomModule/Fcmb/IEmailPhoneService.cs
@@ -16,13 +16,13 @@ namespace CMdm.Services.CustomModule.Fcmb
         /// </summary>
         /// <param name="recordId">recordId identifier</param>
         /// <returns>Vendor</returns>
-        IList<EmailPhone> GetEmailPhonebyIds(int[] recordIds);
+        IList<EmailPhone> GetEmailPhonebyIds(Int64[] recordIds);
         /// <summary>
         /// Gets a Queitem by item reference identifier
         /// </summary>
         /// <param name="recordId">que identifier</param>
         /// <returns>Vendor</returns>
-        EmailPhone GetEmailPhonebyId(int recordId);
+        EmailPhone GetEmailPhonebyId(Int64 recordId);
         /// <summary>
         /// Gets all items
         /// </summary>
